requirements.txt
Flask==2.3.3
openai==1.3.8
sentence-transformers==2.2.2
chromadb==0.4.18
python-dotenv==1.0.0
requests==2.31.0
uuid==1.30